refactor: move Finding value types into homeboy-audit-contract (#8425)
Continues decoupling audit output types from code_audit. Finding (a single
reported issue: kind/severity/file/description/suggestion), its Severity and
FindingConfidence companions, the confidence policy (finding_confidence), the
finding<->HomeboyFinding projection, and the fingerprint/kind-key helpers are the
audit OUTPUT vocabulary — produced by the audit engine, consumed by
refactor/report/CLI. Move them into homeboy-audit-contract as a `finding` module
(the contract gains homeboy-finding + regex deps for the projection).

build_findings (the generator that turns CheckResults into Findings) stays in
code_audit; it now imports Finding/Severity from the contract. code_audit::findings
re-exports the moved items so all existing paths resolve. A duplicate
finding_kind_key wrapper in report.rs is removed in favor of the contract's.

Combined with the AuditFinding move (#8580), the audit finding vocabulary now
lives entirely in the contract, cutting the refactor<->audit coupling that was
wrongly assumed to be by-design.

Verified: contract + core compile clean; 16 contract + 10 findings tests pass;
cargo check --workspace --tests green.

Refs #8425
