+++ crates/homeboy-audit-contract/Cargo.toml
@@ -14,5 +14,7 @@ path = "src/lib.rs"
 
 [dependencies]
 homeboy-engine-primitives = { version = "0.1.0", path = "../homeboy-engine-primitives" }
+homeboy-finding = { version = "0.1.0", path = "../homeboy-finding" }
+regex = "1.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
